drivers: sdram: increase DDR size for KREA
sdram driver limits the size of DDR to 2GB boundary.
This limitation comes from usage of u32 variable to hold the size, and memory
size more than 4GB cannot be passed further. KREA board utilizes 4GB, but
provides only 3GB to the Kernel (1GB is reserved for FPGA). Therefore, for
KREA we raise this boundary to 3GB.

Signed-off-by: Mamta Shukla <mamta.shukla@leica-geosystems.com>